basic -AI/requirements.txt
flask
transformers
torch
speechrecognition
pyttsx3
googlesearch-python